+++ OWC/src/index.ts
@@ -12,6 +12,7 @@ import './dropdown'
 import './tabs'
 import './input'
 import './skeleton'
+import './scroll'
 import { OProgress } from './progress'
 import './progress'
 
@@ -29,6 +30,7 @@ export * from './tabs'
 export * from './input'
 export * from './skeleton'
 export * from './progress'
+export * from './scroll'
 
 // Page-wide theming: <body data-owc-theme="pixel">
 installGlobalThemeStyles()